deadline 3 ure

+++ src/main/java/sorim/f1/slasher/relentless/model/strawpoll/StrawpollPollConfig.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class StrawpollPollConfig {
-    private Integer deadline_at;
+    private Long deadline_at;
     private Boolean is_multiple_choice;
     private Boolean is_private;
